// src/rn-common/lib/components/chart/chart.d.ts
export type Color = string;
export type ValueFormatter = ('largeValue' | 'percent' | 'date') | string | string[];

export type AxisDependency = 'LEFT' | 'RIGHT';
// ...
export interface yAxis extends Axis {
    inverted?: boolean;
    spaceTop?: number;
    spaceBottom?: number;

    position?: 'OUTSIDE_CHART' | 'INSIDE_CHART';

    maxWidth?: number;
    minWidth?: number;

    zeroLine?: {
        enabled?: boolean;
        lineWidth?: number;
        lineColor?: Color;
    };
}

export interface Offsets {
    top?: number;
    left?: number;
    bottom?: number;
    right?: number;
}

export interface ChartBase extends ViewProps {

    drawEntryLabels?: boolean
    animation?: {
        durationX?: number;
        durationY?: number;
        easingX?: EasingType;
        easingY?: EasingType;
    };

    chartBackgroundColor?: Color;
    logEnabled?: boolean;
    noDataText?: string;

    touchEnabled?: boolean;
    dragDecelerationEnabled?: boolean;
    dragDecelerationFrictionCoef?: number;

    highlightPerTapEnabled?: boolean;

    /**
     * 图表文字描述
     */
    chartDescription?: ChartDescription;

    legend?: ChartLegend;

    xAxis?: xAxis;

    marker?: {
        enabled?: boolean;
        digits?: number;
        markerColor?: Color;
        textColor?: Color;
        textSize?: number;
    };

    highlights?: Array<{
        x: number;
        dataSetIndex?: number;
        dataIndex?: number;
        y?: number;
        stackIndex?: number;
    }>;

    onSelect?: (event: ChartSelectEvent) => void;
    onChange?: (event: ChartChangeEvent) => void;
}
export interface BarLineChartBase extends ChartBase {
    maxHighlightDistance?: number;
    drawGridBackground?: boolean;
    gridBackgroundColor?: Color;

    drawBorders?: boolean;
    borderColor?: Color;
    borderWidth?: number;

    minOffset?: number;
    maxVisibleValueCount?: number;
    visibleRange?: {
        x?: {
            min?: number;
            max?: number;
        };
        y?: {
            left?: {
                min?: number;
                max?: number;
            };
            right?: {
                min?: number;
                max?: number;
            };
        };
    };
    autoScaleMinMaxEnabled?: boolean;
    keepPositionOnRotation?: boolean;

    highlightPerDragEnabled?: boolean;

    scaleEnabled?: boolean;
    scaleXEnabled?: boolean;
    scaleYEnabled?: boolean;
    dragEnabled?: boolean;
    pinchZoom?: boolean;
    doubleTapToZoomEnabled?: boolean;

    yAxis?: {
        left?: yAxis;
        right?: yAxis;
    };
    zoom?: {
        scaleX: number;
        scaleY: number;
        xValue: number;
        yValue: number;
        axisDependency?: AxisDependency;
    };

    viewPortOffsets?: Offsets;
    extraOffsets?: Offsets;

    group?: string;
    identifier?: string;
    syncX?: boolean;
    syncY?: boolean;
}

export interface PieRadarChartBase extends ChartBase {
    minOffset?: number;
    rotationEnabled?: boolean;
    rotationAngle?: number;
}
// ...
export interface BarDataset extends Dataset {
    values?: Array<BarValue | number | number[]>;
    config?: BarDatasetConfig;
}

export interface BarData {
    dataSets?: BarDataset[];

    config?: {
        //小数，百分比
        barWidth?: number;
        group?: {
            fromX: number;
            groupSpace: number;
            barSpace: number;
        };
    };
}

export interface BarChartProps extends BarLineChartBase {
    drawValueAboveBar?: boolean;
    drawBarShadow?: boolean;
    highlightFullBarEnabled?: boolean;

    data: BarData;
}
// ...
export interface PieDataset extends Dataset {
    values?: Array<PieValue>;
    label?: string;
    config?: PieDatasetConfig;
}

export interface PieData {
    dataSets?: PieDataset[];
}

export interface PieChartProps extends PieRadarChartBase {
    drawEntryLabels?: boolean;
    usePercentValues?: boolean;

    centerText?: string;
    styledCenterText?: {
        text?: string;
        color?: Color;
        size?: number;
    };

    centerTextRadiusPercent?: number;
    holeRadius?: number;
    holeColor?: Color;
    transparentCircleRadius?: number;
    transparentCircleColor?: Color;

    entryLabelColor?: Color;
    entryLabelTextSize?: number;
    maxAngle?: number;

    data?: PieData;
}